Speed up the Tutel route extraction for k > 2
Building a dense [s, e] mask per route made the gating cost grow with k, so
Tutel lost to the native path once k > 2 on launch-bound configurations. The
top-k columns already name the selected experts, so gather the per-route
index/location/gate values instead, and hand Tutel contiguous int32 routes so
its own per-route casts become no-ops. Resolving the capacity tensor once also
drops a redundant device-to-host sync from the Tutel branch.

Routing output is bit-identical to the previous implementation; measured on
2 GPUs the Tutel/native step-time ratio improves from 1.08 to 1.01 at k=3 and
from 1.16 to 1.04 at k=4, with larger configurations unchanged at ~0.66.

+++ tests/unit/moe/test_moe.py
@@ -306,10 +306,10 @@ class TestTopkGate(DistributedTest):
         reconstructed = torch.zeros_like(combine_weights)
         for indices_s, locations_s, gates_s in zip(indices_, locations_, gates_):
             route_mask = indices_s.ge(0)
-            safe_indices = indices_s.clamp_min(0)
+            safe_indices = indices_s.clamp_min(0).long()
             expert_mask = torch.nn.functional.one_hot(safe_indices, num_classes=num_experts)
             expert_mask *= route_mask.unsqueeze(1)
-            location_mask = torch.nn.functional.one_hot(locations_s, num_classes=int(capacity))
+            location_mask = torch.nn.functional.one_hot(locations_s.long(), num_classes=int(capacity))
             reconstructed += torch.einsum("s,se,sc->sec", gates_s, expert_mask, location_mask)
         torch.testing.assert_close(reconstructed, combine_weights)
